backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
python-multipart==0.0.9
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1
redis==5.0.1
boto3==1.34.48
unstructured[pdf,csv,ppt,epub,markdown,rtf]==0.12.6
# elasticsearch==7.17.10
elasticsearch>=7.17.0,<8.0.0

langchain==0.1.10
langchain-text-splitters==0.0.1
sentence-transformers==2.5.1
pgvector==0.2.4
langchain-community==0.0.25
langchain-core==0.1.29
langchain-openai==0.0.8
tiktoken==0.6.0
openai==1.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 